Fix: Address PR feedback on utility functions and imports
Addressed feedback from PR #5:
- Imported showInfoModal in src/imageUtils.js to resolve runtime errors.
- Refactored handleSaveCharacterToImage in src/imageUtils.js to receive characterData as an argument, decoupling it from direct DOM access. Updated its call in src/index.js accordingly.
- Removed unused 'language' import from src/modalManager.js.
